Added single script to sort both Effects and Generators at the same time. Consolidated CUSTOM scripts into one as they were identical
